RTC operational and communicating with DPI620G
As per RTC specs from STM32 datasheet, both time and after that date has to be read in the sequence for unlocking the values of the registers that have time. Otherwise the RTC does not update.

This issue was present in the earlier commit and has been resolved in this one.

Also functions for getting time and getting date only are also added.